+++ release/release_data_tests.yaml
@@ -394,8 +394,6 @@
 
   cluster:
     byod:
-      runtime_env:
-        - RAY_worker_killing_policy=lifo
       pip:
         - ray[default]
     cluster_compute: "{{scaling}}_all_to_all_compute.yaml"
@@ -412,8 +410,6 @@
 
   cluster:
     byod:
-      runtime_env:
-        - RAY_worker_killing_policy=lifo
       pip:
         - ray[default]
     cluster_compute: dataset/autoscaling_all_to_all_compute.yaml
@@ -436,8 +432,6 @@
 
   cluster:
     byod:
-      runtime_env:
-        - RAY_worker_killing_policy=lifo
       pip:
         - ray[default]
     cluster_compute: "{{scaling}}_all_to_all_compute.yaml"
@@ -453,8 +447,6 @@
 
   cluster:
     byod:
-      runtime_env:
-        - RAY_worker_killing_policy=lifo
       pip:
         - ray[default]
     cluster_compute: dataset/autoscaling_all_to_all_compute.yaml
