+++ release/release_tests.yaml
@@ -2878,9 +2878,6 @@
 
   team: core
   cluster:
-    byod:
-      runtime_env:
-        - RAY_worker_killing_policy=lifo
     cluster_compute: shuffle/shuffle_compute_multi.yaml
 
   run:
@@ -2953,9 +2950,6 @@
 
   team: core
   cluster:
-    byod:
-      runtime_env:
-        - RAY_worker_killing_policy=lifo
     cluster_compute: shuffle/shuffle_compute_autoscaling.yaml
 
   run:
@@ -3121,9 +3115,6 @@
   stable: false
 
   cluster:
-    byod:
-      runtime_env:
-        - RAY_worker_killing_policy=lifo
     cluster_compute: dask_on_ray/dask_on_ray_sort_compute_template.yaml
 
   run:
@@ -3148,9 +3139,6 @@
   team: data
 
   cluster:
-    byod:
-      runtime_env:
-        - RAY_worker_killing_policy=lifo
     cluster_compute: dask_on_ray/dask_on_ray_stress_compute.yaml
 
   run:
@@ -3475,9 +3463,6 @@
   team: core
 
   cluster:
-    byod:
-      runtime_env:
-        - RAY_worker_killing_policy=lifo
     cluster_compute: dask_on_ray/1tb_sort_compute.yaml
 
   run:
