+++ src/ray/common/ray_config_def.h
@@ -90,12 +90,6 @@ RAY_CONFIG(uint64_t, task_failure_entry_ttl_ms, 15 * 60 * 1000)
 /// that is not related to running out of memory. Retries indefinitely if the value is -1.
 RAY_CONFIG(uint64_t, task_oom_retries, -1)
 
-/// The worker killing policy to use, available options are
-/// group_by_owner
-/// lifo
-/// fifo
-RAY_CONFIG(std::string, worker_killing_policy, "group_by_owner")
-
 /// Whether to report placement or regular resource usage for an actor.
 /// Reporting placement may cause the autoscaler to overestimate the resources
 /// required of the cluster, but reporting regular resource may lead to no
